drivers/powerpanel.[ch], man/powerpanel.8: Russ Romano provided enough serial traces for the CyberPower PR2200 so that we could add support for this (binary protocol) to the existing 'powerpanel' driver (text protocol). Which protocol is used, will be detected by the driver automatically.
As of now, this driver obsoletes the 'cpsups' and 'nitram' drivers. Since the protocol that is used by the 'cyberpower' driver is essentially the same as the binary protocol in this driver, that one could be replaced as well. However, from the code it seems that the meaning of some variables are different (specifically the battery charge, temperature, load and possibly the frequency as well). It would be really useful if someone with a CyberPower OPxxxx UPS could run the powerpanel driver and post the first few seconds of debug output, so that we see the full identification string that is sent by the UPS.

git-svn-id: svn://svn.debian.org/nut/trunk@898 72a954d1-e00c-0410-aa02-d9c7bb700a61