Add Belkin F6C1200-UNV (older model) to supported devices.
Apparently this is (internally) identical to the F6C1100-UNV. Not surprisingly, they also share the same productid (1100). It also seems to come in two flavors for the USB version. An older one which is supported through the usbhid-ups driver and a newer one that uses the megatec_usb driver.

git-svn-id: svn://svn.debian.org/nut/trunk@1029 72a954d1-e00c-0410-aa02-d9c7bb700a61