- Flush report buffer after writing new values to the UPS. The data in the UPS may change because of this, so we want to be sure that all subsequent reads, return fresh data. - Use calloc() instead of malloc() where useful. - HIDGetItemString returns (char *) but receives a (unsigned char *) as parameter. - physical_to_logical() would always return an integer, since it casts the result of the value conversion to a long, before returning the result as float.
I remember that we discussed the last item before on the mailinglist, but at that time it was decided not to change this, but I don't remember why exactly. It doesn't seem to do any harm, as we further on (in usbhid-ups) cast the result to a long anyway, when we need is as an index value for the value lookup. This change merely does that later on in the process.


git-svn-id: svn://svn.debian.org/nut/trunk@1055 72a954d1-e00c-0410-aa02-d9c7bb700a61