* drivers/libhid.[ch],libshut.c,libusb.c: - Added a callback function in the libshut_open and libusb_open that is called after the report descriptor is read, but before the device is accepted.
* drivers/megatec.c:
- Applied change above.

This change isn't particularly useful yet. The idea is, that the callback function is located in the driver (not libhid, this is just for compatibility with other drivers that use the HIDOpenDevice), so that it can determine whether this device is acceptable or not. This allows for much more flexibility than doing this in libhid.c or even libusb.c. Even if the UPS hides it's unique identification deep inside some non-standard location, we would still be able to get to it before determining that this is the device we want to (re)connect to. We probably may even be able to check if this is a UPS / Powerdevice, which would allow us to match a lot broader on the ProductID (and maybe even remove the match for that altogether). If the device isn't a UPS / Powerdevice, we would simply reject it in the callback function and libusb_open would try the next device.


git-svn-id: svn://svn.debian.org/nut/trunk@1075 72a954d1-e00c-0410-aa02-d9c7bb700a61