- Removed new "shutdown.restart" variable and changed back to 'shutdown.return'. - Corrected shutdown function (should fallback to 'shutdown.reboot', not 'shutdown.return').
We should rework the variables and commands that deal with shutting down the UPS. It certainly should be much clearer when a command can be used with a command and when not. With the option of adding parameters to commands, the differences between commands and variables are not so clear-cut anymore. Until we have decided how to proceed with this, it is probably best to make do with what we already have.

git-svn-id: svn://svn.debian.org/nut/trunk@1114 72a954d1-e00c-0410-aa02-d9c7bb700a61