Parsing continues if the HID path for a report ends in 0x00000000 (these will not be made available). Commonly used as placeholders.
It looks like (but correct me if I'm wrong) that reports for which the HID path ends in 0x00000000 are placeholders only. By ignoring them already during the parsing of the report descriptor, they will not end up in the list of pData elements in the parsed report descriptor, so we don't have to ignore them later on (with every interrupt report we receive). This saves both time and complexity.

git-svn-id: svn://svn.debian.org/nut/trunk@1210 72a954d1-e00c-0410-aa02-d9c7bb700a61