Use poll() instead of select() to allow larger numbers of clients to connect to the server than FD_SETSIZE allows (maximum number can be specified with MAXCONN in upsd.conf, defaults to FD_SETSIZE).
Although it is possible to dynamically adjust the number of connections allowed, it is intentionally only possible to change this number after a server reload and the storage that is needed is (re)allocated only then. This in order to prevent hard to diagnose memory allocation problems at odd times when the server is running, possibly with nobody around to fix them (we don't want to risk that).



git-svn-id: svn://svn.debian.org/nut/trunk@1376 72a954d1-e00c-0410-aa02-d9c7bb700a61