be a good little client that cleans up after itself
One thing that is not so good about this client, is that it uses 'getpass' to read the password from the command line. This has been marked 'obsolete' by POSIX several years ago already and is generally considered to be not very portable (although so far not many people seem to have complained about us using it). Any suggestions here?

git-svn-id: svn://svn.debian.org/nut/trunk@1412 72a954d1-e00c-0410-aa02-d9c7bb700a61