use connect with timeout to prevent the server declaring us stale (requires neon-0.27.0 or newer)
Having ne_set_connect_timeout() is absolutely needed in order to prevent the server declaring us stale while the driver is waiting for the connect to timeout. Therefor, I have decided to list neon >= 0.27.0 as a requirement and not try to build a driver that doesn't use this function. Apologies for the people that are using older neon libraries, but this will prevent lots of people complaining about staleness warnings.

git-svn-id: svn://svn.debian.org/nut/trunk@1428 72a954d1-e00c-0410-aa02-d9c7bb700a61