use GET_OBJECT page instead of XML_SUMMARY_PAGE to read parameters from the UPS
I'm aware that this will break the driver on the MGE 66102 (Mosaic M) with firmware before version EA. If we ever want to be able to support writing variables, we'll have to do this since many variables are not available through the XML_SUMMARY_PAGE. It is better to do this now that the driver is still flagged 'experimental' (and we still have a small user base for it), then to have to do this later on. And no, I don't think supporting the old (broken) firmware is a good idea, now that a fixed version is available.

git-svn-id: svn://svn.debian.org/nut/trunk@1493 72a954d1-e00c-0410-aa02-d9c7bb700a61