Avoid stack corruption on invalid frame
In the protocol, the frame size can be 8 bytes max. That is why the
frame buffer is only 8 bytes long.
If the frame is corrupted, the size read may be up to 15, which may
corrupt the stack while writing into the frame buffer.
This patch add a simple sanity check on the size to avoid a crash
and enforce protocol limitation.

git-svn-id: svn://svn.debian.org/nut/trunk@3503 72a954d1-e00c-0410-aa02-d9c7bb700a61