Add a tweak for APC Back UPS ES
APC Back UPS ES have a buggy firmware which overflows on ReportID 0x0c,
i.e. UPS.PowerSummary.RemainingCapacity.
This results in battery.charge not being exposed and endless reconnections
on systems with libusb reporting EOVERFLOW. And it results on a failure to init
the driver for  systems with libusb not reporting EOVERFLOW but EIO (i.e. on Windows).

git-svn-id: svn://svn.debian.org/nut/trunk@3723 72a954d1-e00c-0410-aa02-d9c7bb700a61