Fix Buildbot failures on commit r3776
The merge of the libnutclient branch caused a failure of the
'distcheck-light' test target. Manual pages documentation in this branch
uses a mechanism to generate multiple manpages from one source file. This
was however conflicting with a Makefile rule, that requires the generated
file to have the same name as the source file. Applies the same principle
by adding the content of the Header section to the NAME commands list.
Also fix a typo error in the Header section of libnutclient_devices

git-svn-id: svn://svn.debian.org/nut/trunk@3777 72a954d1-e00c-0410-aa02-d9c7bb700a61